Remove schema from LuaAvroDatum type
In the latest versions of the Avro library, the avro_datum_t class
contains a reference to the schema that the value is an instance of.
That means we don't have to keep our own reference to the schema in the
Lua wrapper; we can always get the schema when we need it by calling
avro_datum_get_schema.