“Raw” encoding of values
This patch adds a new function that allows you to encode an Avro value
into an arbitrary memory region.  You can use the new encoded_size()
method to determine how big this memory region should be.  The function
is provided as a top-level function in the "avro" module, so that it's
easier to exclude it from any sandboxes you might want to create; since
it takes in an arbitrary pointer as a parameter, it's definitely not a
safe function to include in sandboxes.