LuaJIT FFI implementation
This patch adds a second implementation of the Avro C library bindings,
written using LuaJIT's FFI bindings.  This will allow LuaJIT's JIT
compiler to optimize traces involving the Avro bindings.  The FFI
implementation is a simple translation of the existing “legacy”
bindings.  Loading in the "avro" module dynamically chooses which
implementation to use, based on whether the "ffi" module is available.