Avro data file functions
This patch adds functions for accessing the Avro C library's data file
API.  The new open() function mimics the built-in Lua io.open()
function.  It's called either as:

   local file = avro.open(filename, "r")

or

   local file = avro.open(filename, "w", writer_schema)

The result is an object with either a read() or write() method, and a
close() method.  The read() method takes in an optional AvroValue
instance; if given, the next value in the file is read into that value.
If not given, a new AvroValue is created.  In both cases, the AvroValue
instance is returned.  The write() method takes in a required AvroValue
instance, which must match the schema given when the file was opened.