New wrapper class for record syntax
This patch separates the raw Lua Avro wrappers from the code that
implements the nice record-like syntax for accessing Avro value
contents.  The public API is, for the most part, identical to what was
there before.  (The only change is the somewhat undocumented raw_value
function, which now returns a raw value.  There's a new counterpart,
wrapped_value, if you want a record-syntax wrapper.)

This new setup will allow us to introduce a new hook system, though,
which will let the user insert custom wrapper objects for certain named
Avro schemas.  We'll use this, for instance, to have the Avro value
accessors automatically instantiate a specialized object for accessing
the contents of an IP address, instead of forcing the user to interact
with it as an Avro fixed.