Explicit freeing of Value objects
The LuaJIT trace compiler currently can't compile calls to a custom C
object's finalizer.  We use a lot of Value objects in some pretty tight
loops, however, so we definitely need those to be compiled down.  This
patch replaces the __gc metamethod with an explicit release() method.
This doesn't need to be called on “subvalues” — the elements in an array
or the fields in a record, for instance.  It only needs to be called on
the top-level values created by one of the various new_value() Lua
methods.  Those methods tend not to be called by user-facing code
anyway, at least in Sawmill.  So we'll need some careful valgrind
testing, but it shouldn't be too bad.