Value:set_from_ast() method
This patch adds a new set_from_ast() method, which allows you to fill in
the contents of an Avro value from a pure Lua value.  For scalars, the
AST object must be a compatible Lua scalar.  For arrays, it should be an
array-like table.  For maps, it should be a table with string keys.  For
records, it should be a table, where the keys are the names of the
record's fields.  For unions, you can provide a raw nil value if the
union contains the null schema; otherwise it should be a one-element
table, where the key is the name of one of the schemas in the union.
(This should remind you of the JSON encoding for Avro values...)