Value:to_json() method
This patch adds a new to_json() method to the Value class.  It does the
same thing as the __tostring metamethod, but allows you to get the
string representation of an Avro value using an explicit method call.
(Which is good in LuaJIT, since its trace compiler doesn't compile down
most metamethod calls yet.)