Allow Value wrappers for scalar iteration results
The iterate() method walks through the elements of an Avro array or map.
Before, it would always extract out the contents of any scalar elements,
returning their Lua equivalents.  You can now pass in "true" as a
parameter to the iterate() method, and that will cause the results to
always be full Value objects, even if they represent Avro scalars.