Customizable wrapper classes
The user of the Lua Avro library can now provide customized wrapper
classes for any named schema, which allow us to provide nicer Lua
interfaces to specific kinds of Avro data.  The custom class from the
previous patch, which implements the nice record-like syntax for
accessing the contents of compound Avro values, is the default wrapper
for any compound schema that doesn't have a custom wrapper provided by
the user.  There is also a default wrapper class for scalars, which
converts the Avro value to its Lua equivalent.

This patch does change the public API.  All of the existing
functionality which could work either on wrapped values or raw values
now only work on raw values.  This includes the resolver classes and
Avro data file classes.  This is needed because there's no longer a
standard interface that each wrapper class must conform to — and we
can't really create one, since we're using built-in Lua values as
wrappers in some cases.