More flexible Schema constructor
The Schema constructor now lets you pass in a raw Lua string that names
one of the primitives types, or an existing Schema object.  (This is in
addition to already accepting a Lua string containing a JSON-encoded
schema description.)

In addition, there's a new avro.schema module which predefines all of
the Avro primitives.  These are included in the top-level avro module,
as well.