Link helper constructors
Our implementation of the link helper constructor is interesting because
it's simpler than you'd expect it to be.  Since we're using Lua
expressions to construct Avro schemas, you'd assume that the innermost
part of an expression would get evaluated first.  If this were the case,
it would be hard to construct link schemas (at least, without adding an
avro_schema_link_set_target function to the Avro C library).  The Avro
spec states that the JSON encoding a schema is processed top-down (or
outermost-first), which is opposite to the expression evaluation
strategy in Lua (or any other non-lazy language).  This means that the
link schema would get created before the link's target gets created.
This would require some kind of post-processing step to match up the
links with their targets.

Luckily, we don't fall into this trip, because of how we use curried
functions to implement the "record" constructor.  The syntax looks like:

  local schema = record "name" { fields }

Internally, "record" is a function that takes in only the name; it then
returns a dynamically-constructed function that takes in the fields
table.  This is key; that gives a small gap where we can save the
(currently empty) record schema into a link targets table, before
evaluating the function that processes the fields table.  End result is
that the Lua expressions are actually evaluated outermost-in, and so we
don't need to do any post-processing of the link targets.