clone() and add_fields() methods
This patch adds a couple of helper methods for working with schemas.
The clone() method creates a new schema object with the same contents.
The add_fields_in_place() method adds additional fields to a record
schema.  The add_fields() method creates a new copy of the record
schema, and then adds additional fields to the schema.