Create value implementations lazily
Our Lua wrappers for an Avro schema have an entry for a value
implementation for the schema, in addition to the avro_schema_t itself.
This lets us create new avro_value_t instances for the schema.  Before,
we would initialize this interface whenever we constructed a schema.
However, this doesn't play well with the Lua schema constructors, since
they modify the schema without modifying the value implementation.

We now create the value implementation lazily, when the first value of
that schema is created.  This ensures that the implementation takes into
account any additions that have been made to the schema by the Lua
schema constructors.