Minor data converions in Value:set_from_ast()
The set_from_ast() method now does some minor data conversions on
scalars, if the type of the Lua parameter doesn't exactly match the
underlying Avro type.  For instance, we'll call tonumber() on the
parameter for the numeric Avro types (int, long, float, and double), and
we'll call tostring() on the parameter for the “stringish” types (bytes,
string, and fixed).

Closes: #30