Schema:size() returns Lua number
The FFI implementation of Schema:size() used to return the size_t result
directly, which on 64-bit systems, resulted in an FFI cdata object that
couldn't be used as a number in many contexts.  We're probably not going
to have many fixed schemas that contain more than 2^32 bytes in them, so
the size() method now always casts its result into a Lua number.