Allow repeated definitions in schema JSON
When constructing a schema from a JSON string, the Avro spec says that
it's okay for the definition of a particular named schema to appear more
than once, as long as all of the definitions are identical.  (The
alternative is for the first occurrence to be the full definition, and
all subsequent uses of the schema to be a named link.)  This patch
updates our JSON parsing code to allow this.