Aborted initial draft of pure Lua values
This patch was the start of a pure-Lua Value class, much like what we
did with the Schema class earlier.  Turns out that we didn't need this
massive of a change for the particular Sawmill bug I was trying to fix;
instead, I could get away with removing a LuaJIT cast.  This still might
be worth revisiting in the future, though.