Use MIT/X11 license
We weren't clear before on what licensing terms the code was released
under.  (The rockspecs even claimed that it was proprietary!)  We're now
explicit about releasing the code under the MIT/X11 license, just like
the rest of Lua.  I've also included a top-level COPYING file for the
dkjson module, which we use an embedded copy of, and which is also
release under the MIT/X11 license.